Add fixes.
