--- config/general.php
@@ -1,3 +1,0 @@
-<?php
-date_default_timezone_set("Europe/Moscow");
-?>
