+++ data/main.php
@@ -1,46 +1,52 @@
 <?php
+$expirationDate = strtotime("tomorrow");
+$timeLeft = $expirationDate - time();
+$hours_left = floor($timeLeft / 3600);
+$minutes_left = floor(($timeLeft % 3600) / 60);
+$expiringMark = ($hours_left <= 1) ? "timer--finishing" : null;
+
 $adverts = [
     [
         "title" => "2014 Rossignol District Snowboard",
         "category" => "Доски и лыжи",
         "price" => "10999",
         "url" => "img/lot-1.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ],
     [
         "title" => "DC Ply Mens 2016/2017 Snowboard",
         "category" => "Доски и лыжи",
         "price" => "159999",
         "url" => "img/lot-2.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ],
     [
         "title" => "Крепления Union Contact Pro 2015 года размер L/XL",
         "category" => "Крепления",
         "price" => "8000",
         "url" => "img/lot-3.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ],
     [
         "title" => "Ботинки для сноуборда DC Mutiny Charocal",
         "category" => "Ботинки",
         "price" => "10999",
         "url" => "img/lot-4.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ],
     [
         "title" => "Куртка для сноуборда DC Mutiny Charocal",
         "category" => "Одежда",
         "price" => "7500",
         "url" => "img/lot-5.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ],
     [
         "title" => "Маска Oakley Canopy",
         "category" => "Разное",
         "price" => "5400",
         "url" => "img/lot-6.jpg",
-        "expirationDate" => strtotime("tomorrow")
+        "expirationDate" => $expirationDate
     ]
 ];
 ?>
