+++ helpers.php
@@ -143,4 +143,10 @@ function include_template($name, array $data = []) {
     return $result;
 }
 
-
+function formatPrice($number) {
+   $result = ceil($number);
+   if ($result >= 1000) {
+       $result = number_format($result, 0, "", " ");
+   }
+   return $result;
+}
