+++ index.php
@@ -1,14 +1,15 @@
 <?php
-require_once "config/general.php";
+require_once "data/main.php";
+require_once "helpers.php";
 require_once "data/general.php";
 require_once "data/layout.php";
-require_once "data/main.php";
-require_once "lib/main.php";
-require_once "helpers.php";
 
 $content = include_template("main.php", [
     "categories" => $categories,
-    "adverts" => $adverts
+    "adverts" => $adverts,
+    "hours_left" => $hours_left,
+    "minutes_left" => $minutes_left,
+    "expiringMark" => $expiringMark
 ]);
 
 $layout = include_template("layout.php", [
